fix(client): restore gemini exports with browser-safe runtime
