+++ packages/client/package.json
@@ -24,6 +24,7 @@
     "build": "tsup && node scripts/patch-dts.mjs",
     "check:pack": "node scripts/validate-pack-artifact.mjs",
     "dev": "tsup --watch",
+    "prepublishOnly": "pnpm build && pnpm check:pack",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
