+++ packages/client/src/models.ts
@@ -1,121 +1,75 @@
+import type { GeminiModel, GeminiModelInfo } from '../../core/src/models/gemini-shared';
+import { GEMINI_MODELS } from '../../core/src/models/gemini-shared';
+
 export * from '../../core/src/models/claude';
 export * from '../../core/src/models/codex';
+export * from '../../core/src/models/gemini-shared';
 
-export type GeminiModel =
-  | 'gemini-3-flash'
-  | 'gemini-3-pro'
-  | 'gemini-2.5-pro'
-  | 'gemini-2.5-flash'
-  | 'gemini-2.5-flash-lite'
-  | 'gemini-2.0-flash'
-  | 'gemini-2.0-flash-lite'
-  | 'gemini-2.0-pro'
-  | 'gemini-2.0-flash-thinking-experimental';
+interface ModelCache {
+  models: GeminiModelInfo[];
+  expiresAt: number;
+}
 
-export const DEFAULT_GEMINI_MODEL: GeminiModel = 'gemini-2.0-flash';
+interface GeminiModelsListResponse {
+  models?: Array<{
+    name?: string;
+    displayName?: string;
+    description?: string;
+    supportedGenerationMethods?: string[];
+    inputTokenLimit?: number;
+    outputTokenLimit?: number;
+  }>;
+}
 
-export const GEMINI_MODELS: Record<
-  GeminiModel,
-  {
-    name: string;
-    description: string;
-    inputPrice: string;
-    outputPrice: string;
-    useCase: string;
+let modelCache: ModelCache | null = null;
+const CACHE_TTL = 24 * 60 * 60 * 1000;
+
+export async function fetchGeminiModels(
+  apiKey?: string,
+  forceRefresh = false
+): Promise<GeminiModelInfo[]> {
+  if (!forceRefresh && modelCache && Date.now() < modelCache.expiresAt) {
+    return modelCache.models;
   }
-> = {
-  'gemini-3-flash': {
-    name: 'Gemini 3 Flash',
-    description: 'Latest Flash model - fast responses with strong capabilities',
-    inputPrice: 'TBD',
-    outputPrice: 'TBD',
-    useCase: 'General coding tasks, fast iteration, great price-to-performance',
-  },
-  'gemini-3-pro': {
-    name: 'Gemini 3 Pro',
-    description: 'Latest and most intelligent model (requires Ultra subscription or waitlist)',
-    inputPrice: 'Premium',
-    outputPrice: 'Premium',
-    useCase: 'Most complex tasks, advanced reasoning, state-of-the-art performance',
-  },
-  'gemini-2.5-pro': {
-    name: 'Gemini 2.5 Pro',
-    description: 'Most capable 2.5 model for complex reasoning and multi-step tasks',
-    inputPrice: 'Higher',
-    outputPrice: 'Higher',
-    useCase: 'Complex refactoring, architecture decisions, advanced debugging',
-  },
-  'gemini-2.5-flash': {
-    name: 'Gemini 2.5 Flash',
-    description: 'Balanced performance and cost for most agentic coding tasks',
-    inputPrice: '$0.30',
-    outputPrice: '$2.50',
-    useCase: 'Feature development, bug fixes, code reviews, testing',
-  },
-  'gemini-2.5-flash-lite': {
-    name: 'Gemini 2.5 Flash-Lite',
-    description: 'Ultra-fast, low-cost model for simple tasks',
-    inputPrice: '$0.10',
-    outputPrice: '$0.40',
-    useCase: 'File search, summaries, simple edits, code formatting',
-  },
-  'gemini-2.0-flash': {
-    name: 'Gemini 2.0 Flash',
-    description: "Google's default model (Jan 2025) - next-gen features with superior speed",
-    inputPrice: '$0.15',
-    outputPrice: '$0.60',
-    useCase: 'General purpose coding, native tool use, 1M token context',
-  },
-  'gemini-2.0-flash-lite': {
-    name: 'Gemini 2.0 Flash-Lite',
-    description: 'Ultra-efficient for simple, high-frequency tasks',
-    inputPrice: '$0.075',
-    outputPrice: '$0.30',
-    useCase: 'Simple edits, quick queries, high-volume operations',
-  },
-  'gemini-2.0-pro': {
-    name: 'Gemini 2.0 Pro',
-    description: 'Advanced reasoning and complex problem solving',
-    inputPrice: '$1.25',
-    outputPrice: '$5.00',
-    useCase: 'Complex architecture, advanced algorithms, deep refactoring',
-  },
-  'gemini-2.0-flash-thinking-experimental': {
-    name: 'Gemini 2.0 Flash Thinking (Experimental)',
-    description: 'Shows detailed reasoning process when responding',
-    inputPrice: '$0.15',
-    outputPrice: '$0.60',
-    useCase: 'Learning, debugging reasoning, understanding AI decision-making',
-  },
-};
-
-const DEFAULT_GEMINI_CONTEXT_LIMIT = 1_048_576;
-
-export const GEMINI_CONTEXT_LIMITS: Record<string, number> = {
-  'gemini-3-flash': 1_048_576,
-  'gemini-3-pro': 1_048_576,
-  'gemini-2.5-pro': 1_048_576,
-  'gemini-2.5-flash': 1_048_576,
-  'gemini-2.5-flash-lite': 1_048_576,
-  'gemini-2.0-flash': 1_048_576,
-  'gemini-2.0-flash-lite': 1_048_576,
-  'gemini-2.0-pro': 1_048_576,
-  'gemini-2.0-flash-thinking-experimental': 1_048_576,
-};
-
-export function getGeminiContextWindowLimit(model?: string): number {
-  if (!model) return DEFAULT_GEMINI_CONTEXT_LIMIT;
-
-  const normalized = model.toLowerCase();
-  if (GEMINI_CONTEXT_LIMITS[normalized]) {
-    return GEMINI_CONTEXT_LIMITS[normalized];
+  if (!apiKey) {
+    throw new Error('API key required for fetching Gemini models');
   }
 
-  for (const [key, limit] of Object.entries(GEMINI_CONTEXT_LIMITS)) {
-    if (normalized.startsWith(`${key}-`)) {
-      return limit;
-    }
+  const response = await fetch(
+    `https://generativelanguage.googleapis.com/v1beta/models?key=${encodeURIComponent(apiKey)}`
+  );
+  if (!response.ok) {
+    throw new Error(`Gemini model fetch failed: ${response.status} ${response.statusText}`);
   }
 
-  return DEFAULT_GEMINI_CONTEXT_LIMIT;
+  const payload = (await response.json()) as GeminiModelsListResponse;
+  const models: GeminiModelInfo[] = (payload.models ?? [])
+    .filter((model) => model.name && model.supportedGenerationMethods?.includes('generateContent'))
+    .map((model) => ({
+      name: model.name!.replace('models/', ''),
+      displayName: model.displayName || model.name!,
+      description: model.description,
+      supportedActions: model.supportedGenerationMethods ?? [],
+      inputTokenLimit: model.inputTokenLimit,
+      outputTokenLimit: model.outputTokenLimit,
+    }));
+
+  modelCache = {
+    models,
+    expiresAt: Date.now() + CACHE_TTL,
+  };
+  return models;
+}
+
+export async function getAvailableGeminiModels(apiKey?: string): Promise<string[]> {
+  try {
+    const dynamicModels = await fetchGeminiModels(apiKey);
+    return dynamicModels.map((model) => model.name);
+  } catch (_error) {
+    return Object.keys(GEMINI_MODELS) as GeminiModel[];
+  }
+}
+
+export function clearGeminiModelCache(): void {
+  modelCache = null;
 }
